Split up flowtypes
