+++ packages/colony-js-contract-client/src/flowtypes/params.js
@@ -1,0 +1,22 @@
+/* @flow */
+
+export type ParamTypes =
+  | 'address'
+  | 'bigNumber'
+  | 'boolean'
+  | 'date'
+  | 'hexString'
+  | 'ipfsHash'
+  | 'number'
+  | 'tokenAddress'
+  | 'string';
+
+// [param name, param type, default value (optional)]
+export type Param = [string, ParamTypes, *];
+export type Params = Array<Param>;
+
+export type ParamTypeDef = {
+  validate: (value: any) => boolean,
+  convertOutput: (value: any) => *,
+  convertInput: (value: any) => *,
+};
